% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31-autoplot-varm.R
\name{autoplot.irf_varm}
\alias{autoplot.irf_varm}
\title{Plotting methods for objects of \code{varm}}
\usage{
\method{autoplot}{irf_varm}(
  object,
  impulse = NULL,
  response = NULL,
  var_names = NULL,
  color = "black",
  linetype = "solid",
  show_zero = FALSE,
  label = "{fcol(impulse, 'red')} to {fcol(response, 'blue')}",
  primary_bands = bands(),
  secondary_bands = sec_bands(),
  ...
)
}
\description{
Plotting methods for objects of \code{varm}
}
\examples{


}
