% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_var.R
\name{gen_var}
\alias{gen_var}
\title{Acoef <- var_ls(abvar::spec(econdata::sw2001\link{,-1}, .endo_lags = 4))$coefficients
tsDyn::VAR.sim(t(Acoef\link{-1,}), n = 200, lag = 4, include = "none")
https://github.com/MatthieuStigler/tsDyn/blob/a6aa2ebc041e448b3eb8598ffd117a83cfe316e5/tsDyn/R/VAR.sim.R}
\usage{
gen_var(Acoef, innov = NULL, varcov = NULL, nsim = 100)
}
\description{
Acoef <- var_ls(abvar::spec(econdata::sw2001\link{,-1}, .endo_lags = 4))$coefficients
tsDyn::VAR.sim(t(Acoef\link{-1,}), n = 200, lag = 4, include = "none")
https://github.com/MatthieuStigler/tsDyn/blob/a6aa2ebc041e448b3eb8598ffd117a83cfe316e5/tsDyn/R/VAR.sim.R
}
