% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{id_triangular}
\alias{id_triangular}
\title{Identification with triangluar decomposition}
\usage{
id_triangular(x, ...)
}
\description{
Identification with triangluar decomposition
}
\examples{
\dontrun{
A <- matrix(c(1, 1, 1, 1, 1, 1, -1, -1, 1, -1, -1, 1, 1, -1, 1, -1), 4)
luDec <- lu(A)
L <- expand(luDec)$L
U <- expand(luDec)$U
p <- expand(luDec)$P
# LU is a row-permuted version of A
L \%*\% U
#Going back to the original identity A = PLU we can recover A
P \%*\% L \%*\% U
}

}
