% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21-irf-varm.R
\name{irf.varm}
\alias{irf.varm}
\title{Impulse response functions}
\usage{
\method{irf}{varm}(
  object,
  horizon = 12,
  nboot = 100,
  boot = boot_spec(fn = boot_inst, nboot = 500),
  id = id_none(),
  shock = shock_scale(),
  ...
)
}
\description{
@export
}
\examples{
\dontrun{



irf(obj_dt, nboot = 100, horizon = 20, id = id_chol(), shock = scale_shock("se")) \%>\% str()
irf(obj_dt, nboot = 100, horizon = 20, id = id_chol(), shock = scale_shock("unit")) \%>\% str()

autoplot(irf(obj_dt, nboot = 100, horizon = 20, id = id_none(), shock = scale_shock("unit")))

autoplot(irf(obj_dt, nboot = 100, horizon = 20, id = id_chol(), shock = scale_shock("unit")))


}
}
