% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf_algorithms.R
\name{irf_algo3}
\alias{irf_algo3}
\title{Impulse Response Function Algorithm 2 (this may be wrong)}
\usage{
irf_algo3(Acomp, B, K, p, h)
}
\description{
Impulse Response Function Algorithm 2 (this may be wrong)
}
