% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psummary.R
\name{psummary.varm}
\alias{psummary.varm}
\title{Pretty Summary}
\usage{
\method{psummary}{varm}(
  object,
  include = c("none", "tstat", "pvalue", "both"),
  digits = max(3L, getOption("digits") - 3L),
  ...
)
}
\description{
Pretty Summary
}
