% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_var.R
\name{sim_var}
\alias{sim_var}
\title{Simulate VAR}
\usage{
sim_var(Acoef, innov = NULL, varcov = NULL, nrep = 200)
}
\description{
Simulate VAR
}
\examples{

obj <- varm(abvar::spec(econdata::sw2001[,-1], .endo_lags = 4))
Acoef <- coefficients(obj)
set.seed(123)
dyn <- tsDyn:::VAR.sim(t(Acoef), n = 200, lag = 4, include = "const")
plot.ts(dyn)
set.seed(123)
ab <- gen_var(Acoef, nrep = 200)
plot.ts(ab)
waldo::compare(dyn, ab)


}
\seealso{
mlvar::simulateVAR  tsDyn::VAR.sim
}
