% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-spec.R
\name{spec}
\alias{spec}
\title{Create a model specification that can fit into multiple models}
\usage{
spec(
  .data,
  .formula = NULL,
  .index = NULL,
  .ordering = NULL,
  .endo_lags = 2,
  .endo_names = NULL,
  .trend = FALSE,
  .factor_names = NULL,
  .tvp = NULL,
  .sv = NULL,
  .exo_names = NULL,
  .exo_lags = NULL
)
}
\arguments{
\item{.data}{df}

\item{.formula}{the formula}

\item{.endo_lags}{number of endogenous lags}

\item{.exo_lags}{number of exogenous lags}
}
\description{
Create a model specification that can fit into multiple models
}
\examples{
spec(yraw)

}
