% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favar.R
\name{factor_extract}
\alias{factor_extract}
\title{Extracts first k principal components from t x n matrix data, loadings
are normalized so that lam'lam/n=I, fac is t x k, lam is n x k}
\usage{
factor_extract(factordata, no_factors = 2)
}
\arguments{
\item{.data}{data from which the factors should be extracted}

\item{nfactors}{number of factors that are going to be extracted}
}
\value{
matrix with the extracted factors
}
\description{
Extracts first k principal components from t x n matrix data, loadings
are normalized so that lam'lam/n=I, fac is t x k, lam is n x k
}
