% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec}
\alias{spec}
\title{Model Specification}
\usage{
spec(.data, .formula, .endo_lags = 2, .exo_lags = 2)
}
\description{
Model Specification
}
\examples{

.data <-
 tibble(
   x = rnorm(100),
   y = rnorm(100),
   z = rnorm(100)
 )

.data \%>\%
  add_trend() \%>\%
  model_spec(x + y ~ 1) \%>\%
}
