% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec}
\alias{spec}
\title{Specification}
\usage{
spec(.data, .formula, .endo_lags = 2, .exo_lags = 2)
}
\arguments{
\item{.data}{df}

\item{.formula}{the formula}

\item{.endo_lags}{of endogenous lags}

\item{.exo_lags}{of exogenous lags}
}
\description{
Specification
}
